User post relationship

+++ app/routers/posts.py
@@ -18,7 +18,7 @@ router = APIRouter(
 # we are returning a list of posts, whereas the response model tries to fit that
 # into the model for one single post as is defined in PostResponse 🤦‍♂️
 # This is why need to import List[] from typing library
-def get_posts(db: Session = Depends(get_db), current_user: int = Depends(get_current_user)):
+def get_posts(db: Session = Depends(get_db), current_user: dict = Depends(get_current_user)):
     posts = db.query(models.Post).all()
 
     return posts
@@ -28,8 +28,7 @@ def get_posts(db: Session = Depends(get_db), current_user: int = Depends(get_cur
 def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db),
                  current_user: int = Depends(get_current_user)):
     # ALT: new_post = models.Post(title=post.title, content=post.content, published=post.published)
-    print(current_user.email, current_user.id)
-    new_post = models.Post(**post.model_dump())
+    new_post = models.Post(**post.model_dump(), owner_id=current_user.id)
     db.add(new_post)
     db.commit()
     db.refresh(new_post)
@@ -38,7 +37,7 @@ def create_posts(post: schemas.PostCreate, db: Session = Depends(get_db),
 
 
 @router.get('/{post_id}', response_model=schemas.PostResponse)
-def get_post(post_id: int, db: Session = Depends(get_db), current_user: int = Depends(get_current_user)):
+def get_post(post_id: int, db: Session = Depends(get_db), current_user: dict = Depends(get_current_user)):
     fetched_post = db.query(models.Post).get(
         {"id": post_id})  # get finds via PK
     if not fetched_post:
@@ -51,7 +50,8 @@ def get_post(post_id: int, db: Session = Depends(get_db), current_user: int = De
 
 
 @router.delete('/{post_id}', status_code=status.HTTP_204_NO_CONTENT)
-def delete_post(post_id: int, db: Session = Depends(get_db), current_user: int = Depends(get_current_user)):
+def delete_post(post_id: int, db: Session = Depends(get_db),
+                current_user: dict = Depends(get_current_user)):
     post_query = db.query(models.Post).filter(
         models.Post.id == post_id)
     post_to_be_deleted = post_query.first()
@@ -60,19 +60,30 @@ def delete_post(post_id: int, db: Session = Depends(get_db), current_user: int =
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with post id: {post_id} not found."
         )
+    if post_to_be_deleted.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail='Cannot delete a post that you do not own.'
+        )
     #! The query needs to be "deleted" not the post itself 🤔
     post_query.delete(synchronize_session=False)
     db.commit()
 
 
 @router.put('/{post_id}', response_model=schemas.PostResponse)
-def update_post(post_id: int, post: schemas.PostUpdate, db: Session = Depends(get_db), current_user: int = Depends(get_current_user)):
+def update_post(post_id: int, post: schemas.PostUpdate, db: Session = Depends(get_db),
+                current_user: dict = Depends(get_current_user)):
     post_query = db.query(models.Post).filter(models.Post.id == post_id)
     post_to_be_updated = post_query.first()
     if not post_to_be_updated:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Post with post id: {post_id} not found."
+        )
+    if post_to_be_updated.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail='Cannot update a post that you do not own.'
         )
     #! The query needs to be "updated" not the post itself 🤔
     post_query.update(post.model_dump(), synchronize_session=False)
